+++ block/action_test.go
@@ -1,4 +1,4 @@
-package action
+package block
 
 import "testing"
