+++ block/block_input.go
@@ -1,14 +1,13 @@
-package input
+package block
 
 import (
-	"github.com/jeremyforan/go-flocks-of-blocks/block"
 	"github.com/jeremyforan/go-flocks-of-blocks/common"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/compositiontext"
 	"github.com/jeremyforan/go-flocks-of-blocks/element"
 )
 
 type Input struct {
-	slackType block.BlockType
+	slackType BlockType
 	label     compositiontext.CompositionText
 	element   element.InputElement
 
@@ -29,7 +28,7 @@ type inputOptional struct {
 
 func NewInput(label string, element element.InputElement) Input {
 	return Input{
-		slackType: block.Input,
+		slackType: InputBlock,
 		label:     compositiontext.NewPlainText(label),
 		element:   element,
 	}
@@ -138,4 +137,56 @@ func (i Input) Render() string {
 
 type InputType interface {
 	Input()
+}
+
+func (i Input) SetLabel(label string) Input {
+	i.setLabel(label)
+	return i
+}
+
+// EnableDispatchAction bool as chain method
+func (i Input) EnableDispatchAction() Input {
+	i.setDispatchAction(true)
+	return i
+}
+
+// DisableDispatchAction bool as chain method
+func (i Input) DisableDispatchAction() Input {
+	i.setDispatchAction(false)
+	return i
+}
+
+// SetBlockID BlockID string as chain method
+func (i Input) SetBlockID(blockID string) Input {
+	i.setBlockID(blockID)
+	return i
+}
+
+// RemoveBlockID string as chain method
+func (i Input) RemoveBlockID() Input {
+	i.removeBlockID()
+	return i
+}
+
+// SetHint as chain method
+func (i Input) SetHint(s string) Input {
+	i.setHint(s)
+	return i
+}
+
+// RemoveHint as chain method
+func (i Input) RemoveHint() Input {
+	i.removeHint()
+	return i
+}
+
+// SetSlackOptional as chain method
+func (i Input) MakeOptional() Input {
+	i.setSlackOptional()
+	return i
+}
+
+func (i Input) RemoveOptional() Input {
+	i.removeSlackOptional()
+	return i
 }
