+++ block/file_test.go
@@ -1,4 +1,4 @@
-package file
+package block
 
 import (
 	"fmt"
