+++ block/header_test.go
@@ -1,4 +1,4 @@
-package header
+package block
 
 import (
 	"fmt"
