+++ block/image_test.go
@@ -1,4 +1,4 @@
-package image
+package block
 
 import (
 	"fmt"
