+++ block/input_test.go
@@ -1,4 +1,4 @@
-package input
+package block
 
 //import (
 //	"flocksofblocks/element/checkboxes"
