+++ element/multiselectmenu/multi_select_menu_with_external_data_source.go
@@ -1,7 +1,7 @@
 package multiselectmenu
 
 import (
-	"github.com/jeremyforan/go-flocks-of-blocks/block/section"
+	"github.com/jeremyforan/go-flocks-of-blocks/block"
 	"github.com/jeremyforan/go-flocks-of-blocks/common"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/compositiontext"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/confirmationdialog"
@@ -294,7 +294,7 @@ func (m MultiSelectMenuWithExternalDataSource) Render() string {
 	return common.Pretty(raw)
 }
 
-func (m MultiSelectMenuWithExternalDataSource) Section() section.Section {
-	s := section.NewSection("newSection").AddAccessory(m)
+func (m MultiSelectMenuWithExternalDataSource) Section() block.Section {
+	s := block.NewSection("newSection").AddAccessory(m)
 	return s
 }
