+++ element/multiselectmenu/multi_select_menu_with_static_options.go
@@ -1,7 +1,7 @@
 package multiselectmenu
 
 import (
-	"github.com/jeremyforan/go-flocks-of-blocks/block/section"
+	"github.com/jeremyforan/go-flocks-of-blocks/block"
 	"github.com/jeremyforan/go-flocks-of-blocks/common"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/compositiontext"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/confirmationdialog"
@@ -343,7 +343,7 @@ func (m MultiSelectMenuWithStaticOption) Render() string {
 func (m MultiSelectMenuWithStaticOption) ElementRender() {}
 
 // Section public section block
-func (m MultiSelectMenuWithStaticOption) Section() section.Section {
-	s := section.NewSection("newSection").AddAccessory(m)
+func (m MultiSelectMenuWithStaticOption) Section() block.Section {
+	s := block.NewSection("newSection").AddAccessory(m)
 	return s
 }
