+++ element/selectmenu/select_menu_with_external_data_source.go
@@ -1,7 +1,7 @@
 package selectmenu
 
 import (
-	"github.com/jeremyforan/go-flocks-of-blocks/block/section"
+	"github.com/jeremyforan/go-flocks-of-blocks/block"
 	"github.com/jeremyforan/go-flocks-of-blocks/common"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/compositiontext"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/confirmationdialog"
@@ -251,7 +251,7 @@ func (m SelectMenuWithExternalDataSource) Render() string {
 	return common.Pretty(raw)
 }
 
-func (m SelectMenuWithExternalDataSource) Section() section.Section {
-	s := section.NewSection("newSection").AddAccessory(m)
+func (m SelectMenuWithExternalDataSource) Section() block.Section {
+	s := block.NewSection("newSection").AddAccessory(m)
 	return s
 }
