+++ element/selectmenu/select_menu_with_public_channels_select.go
@@ -1,7 +1,7 @@
 package selectmenu
 
 import (
-	"github.com/jeremyforan/go-flocks-of-blocks/block/section"
+	"github.com/jeremyforan/go-flocks-of-blocks/block"
 	"github.com/jeremyforan/go-flocks-of-blocks/common"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/compositiontext"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/confirmationdialog"
@@ -257,7 +257,7 @@ func (m SelectMenuWithPublicChannelsSelect) Render() string {
 	return common.Pretty(raw)
 }
 
-func (m SelectMenuWithPublicChannelsSelect) Section() section.Section {
-	s := section.NewSection("newSection").AddAccessory(m)
+func (m SelectMenuWithPublicChannelsSelect) Section() block.Section {
+	s := block.NewSection("newSection").AddAccessory(m)
 	return s
 }
