+++ element/selectmenu/select_menu_with_static_options.go
@@ -1,7 +1,7 @@
 package selectmenu
 
 import (
-	"github.com/jeremyforan/go-flocks-of-blocks/block/section"
+	"github.com/jeremyforan/go-flocks-of-blocks/block"
 	"github.com/jeremyforan/go-flocks-of-blocks/common"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/compositiontext"
 	"github.com/jeremyforan/go-flocks-of-blocks/composition/confirmationdialog"
@@ -311,7 +311,7 @@ func (m SelectMenuWithStaticOption) Render() string {
 func (m SelectMenuWithStaticOption) ElementRender() {}
 
 // SectionBlock public section block
-func (m SelectMenuWithStaticOption) Section() section.Section {
-	s := section.NewSection("newSection").AddAccessory(m)
+func (m SelectMenuWithStaticOption) Section() block.Section {
+	s := block.NewSection("newSection").AddAccessory(m)
 	return s
 }
